create logger mixin
